binary

+++ search_algorithms/1-binary.c
@@ -16,7 +16,7 @@ int binary_search(int *array, size_t size, int value) {
     while (low <= high) {
         size_t mid = (low + high) / 2;
 
-        printf("Searching in array[%zu:%zu] = {", low, high); 
+        printf("Searching in array[%u:%u] = {", (unsigned)low, (unsigned)high); 
         for (i = low; i <= high; i++) {
             printf("%d", array[i]);
             if (i < high) {
